Fix restoring the mini video window restarting the stream
Reason was that the mini-window and the maximized one use two
different MediaSource components (primaryMediaSource and
backgroundMediaSource). Thus, maximizing it got the player destroyed and
recreated, restarting the stream.

Fix is to use a single MediaSource component (merging the "primary"
and "background" mediasource ref into one "media"), and add a property
indicating background usage or not.